+++ brikar-error-response-parser/src/main/java/com/truward/brikar/error/parser/RestErrorParser.java
@@ -19,7 +19,7 @@ import java.io.IOException;
 public final class RestErrorParser {
   private RestErrorParser() {}
 
-  public static ErrorModel.ErrorV1 parseError(HttpStatusCodeException exception) throws IOException {
+  public static ErrorModel.ErrorV2 parseError(HttpStatusCodeException exception) throws IOException {
     final MediaType contentType = exception.getResponseHeaders().getContentType();
     if (contentType == null) {
       throw new IOException("Missing content type in the response headers");
@@ -27,13 +27,13 @@ public final class RestErrorParser {
 
     try (final ByteArrayInputStream bais = new ByteArrayInputStream(exception.getResponseBodyAsByteArray())) {
       if (ProtobufHttpConstants.PROTOBUF_MEDIA_TYPE.isCompatibleWith(contentType)) {
-        return ErrorModel.ErrorResponseV1.parseDelimitedFrom(bais).getError();
+        return ErrorModel.ErrorResponseV2.parseDelimitedFrom(bais).getError();
       }
 
       if (MediaType.APPLICATION_JSON.isCompatibleWith(contentType)) {
         final JsonFactory jsonFactory = new JsonFactory();
         try (final JsonParser jp = jsonFactory.createParser(bais)) {
-          return ProtobufJacksonUtil.readJson(ErrorModel.ErrorResponseV1.class, jp).getError();
+          return ProtobufJacksonUtil.readJson(ErrorModel.ErrorResponseV2.class, jp).getError();
         }
       }