+++ brikar-server/src/main/java/com/truward/brikar/server/controller/DefaultRestExceptionHandler.java
@@ -5,12 +5,10 @@ import com.truward.brikar.error.RestErrors;
 import com.truward.brikar.error.model.ErrorModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 /**
@@ -30,22 +28,22 @@ public interface DefaultRestExceptionHandler {
 
   @ExceptionHandler(HttpRestErrorException.class)
   @ResponseBody
-  default ResponseEntity<ErrorModel.ErrorResponseV1> restErrorException(HttpRestErrorException e) {
-    final ErrorModel.ErrorResponseV1 body = ErrorModel.ErrorResponseV1.newBuilder().setError(e.getError()).build();
+  default ResponseEntity<ErrorModel.ErrorResponseV2> restErrorException(HttpRestErrorException e) {
+    final ErrorModel.ErrorResponseV2 body = ErrorModel.ErrorResponseV2.newBuilder().setError(e.getError()).build();
     return new ResponseEntity<>(body, HttpStatus.valueOf(e.getStatusCode()));
   }
 
   @ExceptionHandler({IllegalArgumentException.class, UnsupportedOperationException.class})
   @ResponseStatus(HttpStatus.BAD_REQUEST)
   @ResponseBody
-  default ErrorModel.ErrorResponseV1 illegalArgumentOrUnsupportedException(Exception e) {
-    return RestErrors.fromThrowable(e);
+  default ErrorModel.ErrorResponseV2 illegalArgumentOrUnsupportedException(Exception e) {
+    return ErrorModel.ErrorResponseV2.newBuilder().setError(RestErrors.fromThrowable(e)).build();
   }
 
   @ExceptionHandler(Throwable.class)
   @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
   @ResponseBody
-  default ErrorModel.ErrorResponseV1 uncategorizedException(Throwable e) {
-    return RestErrors.fromThrowable(e);
+  default ErrorModel.ErrorResponseV2 uncategorizedException(Throwable e) {
+    return ErrorModel.ErrorResponseV2.newBuilder().setError(RestErrors.fromThrowable(e)).build();
   }
 }